Migrate optimization configurations

+++ src/Configurations/chargingconfiguration.h
@@ -16,7 +16,7 @@ class ChargingConfiguration : public QObject
     Q_PROPERTY(int optimizationMode READ optimizationMode WRITE setOptimizationMode NOTIFY optimizationModeChanged)
     Q_PROPERTY(QUuid uniqueIdentifier READ uniqueIdentifier WRITE setUniqueIdentifier NOTIFY uniqueIdentifierChanged)
     Q_PROPERTY(bool controllableLocalSystem READ controllableLocalSystem WRITE setControllableLocalSystem NOTIFY controllableLocalSystemChanged)
-    Q_PROPERTY(float priceThreshold READ priceThreshold WRITE setPriceThreshold NOTIFY priceThresholdChanged)
+    Q_PROPERTY(double priceThreshold READ priceThreshold WRITE setPriceThreshold NOTIFY priceThresholdChanged)
     Q_PROPERTY(bool relativePriceEnabled READ relativePriceEnabled WRITE setRelativePriceEnabled NOTIFY relativePriceEnabledChanged)
     /*!
      * \brief JSON-serialisierter Wochenzeitplan für den zeitgesteuerten Lademodus (TIME_CONTROLLED).
@@ -97,8 +97,8 @@ public:
     bool controllableLocalSystem() const;
     void setControllableLocalSystem(bool controllableLocalSystem);
 
-    float priceThreshold() const;
-    void setPriceThreshold(float priceThreshold);
+    double priceThreshold() const;
+    void setPriceThreshold(double priceThreshold);
 
     bool relativePriceEnabled() const;
     void setRelativePriceEnabled(bool relativePriceEnabled);
@@ -123,7 +123,7 @@ signals:
     void optimizationModeChanged(int optimizationMode);
     void uniqueIdentifierChanged(QUuid uniqueIdentifier);
     void controllableLocalSystemChanged(bool controllableLocalSystem);
-    void priceThresholdChanged(float priceThreshold);
+    void priceThresholdChanged(double priceThreshold);
     void relativePriceEnabledChanged(bool relativePriceEnabled);
     void chargingScheduleChanged(const QString &chargingSchedule);
     void desiredPhaseCountChanged(uint desiredPhaseCount);
@@ -139,7 +139,7 @@ private:
     int m_optimizationMode = 0;
     QUuid m_uniqueIdentifier = QUuid("2e2d25c5-57c7-419a-b294-881f11ed01c4");
     bool m_controllableLocalSystem = false;
-    float m_priceThreshold = 0.0;
+    double m_priceThreshold = 0.0;
     bool m_relativePriceEnabled = false;
     QString m_chargingSchedule = "";
     uint m_desiredPhaseCount = ThreePhase;
